+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# next
+
 # [0.2.0.0] -- 2025-05-02
 
 [0.2.0.0]: https://github.com/GaloisInc/oughta/releases/tag/v0.2.0.0