+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# next
+# [0.2.0.0] -- 2025-05-02
+
+[0.2.0.0]: https://github.com/GaloisInc/oughta/releases/tag/v0.2.0.0
 
 - Add hooks, `Lua` actions that run before or after Lua programs. To upgrade
   from v0.1, just add `defaultHooks` as the first argument to `check{,'}`.